Fix bug in passing inertial measurements as long instead of double

+++ RCCore/RCCorvisManagerFactory.h
@@ -16,8 +16,8 @@
 - (void)stopPlugins;
 - (BOOL)isPluginsStarted;
 - (void)receiveVideoFrame:(unsigned char*)pixel withWidth:(uint32_t)width withHeight:(uint32_t)height withTimestamp:(CMTime)timestamp;
-- (void)receiveAccelerometerData:(uint64_t)timestamp withX:(long)x withY:(long)y withZ:(long)z;
-- (void)receiveGyroData:(uint64_t)timestamp withX:(long)x withY:(long)y withZ:(long)z;
+- (void)receiveAccelerometerData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z;
+- (void)receiveGyroData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z;
 
 @end
